sync

+++ bin/installers/conda.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/env bash
 
+[ -d "$HOME/.miniconda" ] && return 0
 [ -d "$HOME/miniconda3" ] && return 0
+[ -d "$HOME/.miniconda3" ] && return 0
 [ -d "$HOME/anaconda" ] && return 0
 
 wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda.sh
-bash ~/miniconda.sh -b -p "$HOME/miniconda3"
+bash ~/miniconda.sh -b -p "$HOME/.miniconda"
 rm -f "$HOME/miniconda.sh"
